Bug fix include rel paths (#28)
* Fix an inconsistency with Nastran when following include statements.

If an INCLUDE statement only specifies a filename (i.e. INCLUDE 'file.bdf'), Nastran treats file.bdf as if it's in the same directory as the file with the INCLUDE statement.

If an INCLUDE statement includes path components (i.e. INCLUDE 'dir1/file.bdf' or INCLUDE '../file.bdf') Nastran treats the path as if it's relative to the directory containing the root Nastran input deck.

Previously, rdcards and rdsets always used the second method.  This was fixed.

* Make a copy of kwargs dictionary before recursive call.

* Correct docstring.

---------

Co-authored-by: Josh Ayers <joayers@relativityspace.com>